Got the awsec2 provider to work properly with a ssh session

+++ providers/awsec2.py
@@ -19,9 +19,11 @@ along with loadr.  If not, see <http://www.gnu.org/licenses/>.
 """
 
 import boto3
+import paramiko
 import sys
 
-from paramiko import SSHClient
+from io import StringIO
+from time import sleep
 
 
 class Awsec2:
@@ -37,13 +39,36 @@ class Awsec2:
         self.ec2 = self.session.resource('ec2')
         self.config = config
         self.keypair = self.create_keypair()
+        self.securitygroup = self.create_securitygroup()
         self.instances = []
 
     def create_keypair(self):
         keypair = self.ec2.KeyPair('loadr')
         keypair.delete()
-
         return self.ec2.create_key_pair(KeyName='loadr')
+
+    def create_securitygroup(self):
+        sg = None
+
+        for vpc in self.ec2.vpcs.all():
+            if sg is not None:
+                break
+
+            for group in vpc.security_groups.all():
+                if group.group_name == 'loadr':
+                    sg = group
+                    break
+
+        if sg is None:
+            sg = self.ec2.create_security_group(GroupName='loadr',
+                                                Description='loadr ssh access')
+            sg.authorize_ingress(IpProtocol='tcp',
+                                 CidrIp='0.0.0.0/0',
+                                 FromPort=22,
+                                 ToPort=22)
+
+        return sg
+
 
     def create_instances(self, instances):
         self.instances = self.ec2.create_instances(
@@ -52,22 +77,52 @@ class Awsec2:
                             MinCount=instances,
                             MaxCount=instances,
                             UserData=self.bootscript,
-                            KeyName='loadr')
+                            KeyName=self.keypair.name,
+                            SecurityGroupIds=[self.securitygroup.id])
         [i.wait_until_running() for i in self.instances]
+        [i.load() for i in self.instances]
 
     def remove_instances(self):
-        [i.terminate() for i in self.instances]
-        self.keypair.delete()
+        if self.instances is not None:
+            [i.terminate() for i in self.instances]
+            [i.wait_until_terminated() for i in self.instances]
+            self.instances = None
 
-    def run_worker(self, requests):
+    def run_worker(self, requests, writer):
         for i in self.instances:
-            client = SSHClient()
-            client.connect(i.public_ip_address, pkey=self.keypair.key_material)
-            stdin, stdout, stderr = client.exec_command('ls -lh')
-            print(stdin)
-            print(stdout)
-            print(stderr)
+            client = paramiko.SSHClient()
+            client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+
+            keyfile = StringIO(self.keypair.key_material)
+            key = paramiko.RSAKey.from_private_key(keyfile)
+
+            for x in range(3):
+                try:
+                    client.connect(i.public_dns_name,
+                                   username='ec2-user',
+                                   pkey=key,
+                                   look_for_keys=False,
+                                   timeout=60)
+                    stdin, stdout, stderr = client.exec_command('ls -lah')
+
+                    for l in stdout:
+                        writer.write(l)
+
+                    stdin.close()
+                    stdout.close()
+                    stderr.close()
+                    break
+                except:
+                    sleep(60)
+
+            client.close()
+            keyfile.close()
 
     def shutdown(self):
-        # self.ec2.meta.client.close()
-        pass
+        if self.keypair is not None:
+            self.keypair.delete()
+            self.keypair = None
+
+        if self.securitygroup is not None:
+            self.securitygroup.delete()
+            self.securitygroup = None
